make group row component
